Rename pkg to jcucumberng.api.properties
