+++ src/main/java/project/hooks/ScenarioHook.java
@@ -12,7 +12,7 @@ import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import jcucumberng.api.browser.BrowserFactory;
-import jcucumberng.api.props.PropsLoader;
+import jcucumberng.api.properties.PropsLoader;
 import jcucumberng.api.selenium.Selenium;
 import jcucumberng.api.utils.SystemUtil;
