+++ src/main/java/project/hooks/StepHook.java
@@ -1,7 +1,7 @@
 package project.hooks;
 
 import cucumber.api.java.AfterStep;
-import jcucumberng.api.props.PropsLoader;
+import jcucumberng.api.properties.PropsLoader;
 import jcucumberng.api.selenium.Selenium;
 
 public class StepHook {
